Altered js, removed php, and updated contact section
changed the js so that the api link uses session storage, and only has to make one api call for an array of projects. Removed PHP as it does not run on github pages. Updated contact section so that it does not rely on php form, and updated css to make it look better.